% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_311_noise_data}
\alias{get_311_noise_data}
\title{Load 311 Noise Complaint Data}
\usage{
get_311_noise_data(online = FALSE, package = "NycNoiseViz")
}
\arguments{
\item{online}{load data from the online data API}
}
\value{
311 noise complaint data
}
\description{
Load 311 Noise Complaint Data
}
