% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_facebook_data.R
\name{get_facebook_data}
\alias{get_facebook_data}
\title{Get Facebook Data}
\usage{
get_facebook_data()
}
\value{
scraped and processed Facebook data
}
\description{
This function gets listings from public Facebook groups.
}
