% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_nyc_neighborhoods}
\alias{get_nyc_neighborhoods}
\title{Get NYC Neighborhoods}
\usage{
get_nyc_neighborhoods(tidy = FALSE)
}
\arguments{
\item{tidy}{option for `tidy`}
}
\value{
NYC neighbordhoods
}
\description{
Get NYC Neighborhoods
}
