% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{prc_311_noise_data}
\alias{prc_311_noise_data}
\title{Process Noise Data}
\usage{
prc_311_noise_data(noise_data)
}
\arguments{
\item{noise_data}{loaded noise data}
}
\value{
processed noise data
}
\description{
Process Noise Data
}
