% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_facebook_data.R
\name{prc_facebook_data}
\alias{prc_facebook_data}
\title{Process Facebook data}
\usage{
prc_facebook_data(fb_data)
}
\arguments{
\item{fb_data}{scrapped Facebook data}
}
\value{
processed Facebook data
}
\description{
Process Facebook data
}
