% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reddit_data.R
\name{prc_reddit_data}
\alias{prc_reddit_data}
\title{Process Reddit data}
\usage{
prc_reddit_data(reddit_data)
}
\arguments{
\item{reddit_data}{scrapped Reddit data}
}
\value{
processed reddit data
}
\description{
Process Reddit data
}
