% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{prep_api_key}
\alias{prep_api_key}
\title{Prepare API Key}
\usage{
prep_api_key(package = "NycNoiseViz")
}
\arguments{
\item{package}{default to `NycNoiseViz`.}
}
\value{
api key
}
\description{
This function prepares an API key used in mapping.
}
