% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_facebook_data.R
\name{scrape_facebook_data}
\alias{scrape_facebook_data}
\title{Scrape Facebook Data}
\usage{
scrape_facebook_data(date_time_str)
}
\arguments{
\item{date_time_str}{a unique date time string}
}
\value{
scraped data from the NYCapartments subreddit
}
\description{
This function scraped housing and/or apartment listing
from **public** Facebook groups.
}
