% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reddit_data.R
\name{scrape_reddit_data}
\alias{scrape_reddit_data}
\title{Scrape Reddit Data}
\usage{
scrape_reddit_data(date_time_str)
}
\arguments{
\item{date_time_str}{a unique date time string}
}
\value{
scraped data from the NYCapartments subreddit
}
\description{
Scrape Reddit Data
}
